refactor to arduino pro mega
I used the wrong part because i was trying to recreate the circuit
from memory. Now that this has been rectified i can start finding
a way to generate some output that may be used to program the
router. As it stands the challenge here seems to generate DXF code
from either the Gerber Export or the SVG export. I plan on doing
this in a somewhat reusable way since i haven't been able to find
a freeish software that even comes close to what i seem to need.